Refactor CapacityGraph
