+++ src/client/containers/Home/components/CapacityGraph.tsx
@@ -8,6 +8,12 @@ import {
 } from "victory";
 
 import { CapacityZoomChart } from "./";
+import { formatBytes, formatData, formatTime } from "./utils";
+
+const AXIS: { x: string; y: string } = {
+  x: "date",
+  y: "gbps"
+};
 
 interface CapacityGraphProps {
   bandwidthData: {
@@ -19,45 +25,6 @@ interface CapacityGraphProps {
 interface CapacityGraphState {
   selectedDomain: undefined | DomainPropType;
 }
-
-const Months: string[] = [
-  "Jan",
-  "Feb",
-  "Mar",
-  "Apr",
-  "May",
-  "Jun",
-  "Jul",
-  "Aug",
-  "Sep",
-  "Oct",
-  "Nov",
-  "Dec"
-];
-
-const AXIS: { x: string; y: string } = {
-  x: "date",
-  y: "gbps"
-};
-
-type formatDataType = (
-  data: number[][]
-) => Array<{
-  date: Date;
-  gbps: number;
-}>;
-
-const formatData: formatDataType = (data: number[][]) => {
-  const result = [];
-  for (let i = 0; i < data.length; i += 1) {
-    const newData = {
-      date: new Date(data[i][0]),
-      gbps: data[i][1]
-    };
-    result[i] = newData;
-  }
-  return result;
-};
 
 class CapacityGraph extends Component<CapacityGraphProps, CapacityGraphState> {
   state: CapacityGraphState = {
@@ -76,9 +43,6 @@ class CapacityGraph extends Component<CapacityGraphProps, CapacityGraphState> {
     const { cdn } = this.props.bandwidthData;
     const { selectedDomain } = this.state;
 
-    const axisFormat = (date: number) =>
-      `${new Date(date).getDate()}. ${Months[new Date(date).getMonth()]}`;
-
     const data = formatData(cdn);
     return (
       <>
@@ -96,12 +60,12 @@ class CapacityGraph extends Component<CapacityGraphProps, CapacityGraphState> {
         >
           <VictoryAxis
             // tslint:disable-next-line
-            tickFormat={axisFormat}
+            tickFormat={formatTime}
           />
           <VictoryAxis
             dependentAxis
             // tslint:disable-next-line
-            tickFormat={bytes => `${(bytes / 1073741824).toFixed(1)}\nGbps`}
+            tickFormat={formatBytes}
           />
           <VictoryLine data={data} x={AXIS.x} y={AXIS.y} />
         </VictoryChart>
@@ -111,7 +75,7 @@ class CapacityGraph extends Component<CapacityGraphProps, CapacityGraphState> {
           y={AXIS.y}
           selectedDomain={selectedDomain}
           onBrush={this.handleBrush}
-          tickFormat={axisFormat}
+          tickFormat={formatTime}
         />
       </>
     );
